Move common bonuses code to common folder
