+++ ts/features/bonus/bonusVacanze/screens/eligibility/TimeoutEligibilityCheckInfoScreen.tsx
@@ -3,7 +3,7 @@ import * as React from "react";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import I18n from "../../../../../i18n";
-import { BaseTimeoutScreen } from "../../components/BaseTimeoutScreen";
+import { BaseTimeoutScreen } from "../../../common/BaseTimeoutScreen";
 import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
 import { cancelBonusVacanzeRequest } from "../../store/actions/bonusVacanze";
