+++ ts/features/wallet/onboarding/cobadge/screens/search/ko/CoBadgeKoNotFound.tsx
@@ -19,7 +19,7 @@ import {
 } from "../../../../../../bonus/bonusVacanze/components/buttons/ButtonConfigurations";
 import { FooterStackButton } from "../../../../../../bonus/bonusVacanze/components/buttons/FooterStackButtons";
 import { useHardwareBackButton } from "../../../../../../../hooks/useHardwareBackButton";
-import TosBonusComponent from "../../../../../../bonus/bonusVacanze/components/TosBonusComponent";
+import TosBonusComponent from "../../../../../../bonus/common/components/TosBonusComponent";
 import { walletAddCoBadgeCancel } from "../../../store/actions";
 
 export type Props = ReturnType<typeof mapDispatchToProps> &
